+++ Chess/ChessUtilities/ChessUtilities/src/ChessUtilities/Uci/UciInterpreter.cpp
@@ -55,15 +55,13 @@ UciInterpreter::InfoDetails UciInterpreter::createInfoDetailsFromInfoTokens(stri
         } else if (isACommonParameter(token)) {
             infoDetails.commonParameterNameAndValue.emplace_back(token, tokens.at(++i));
         } else if ("multipv" == token) {
-            infoDetails.multipv = convertStringToNumber<unsigned int>(tokens.at(++i));
+            infoDetails.multipv = convertStringToNumber<int>(tokens.at(++i));
         } else if ("cp" == token) {
             infoDetails.scoreInCentipawns = convertStringToNumber<int>(tokens.at(++i));
-        } else if ("mate" == token) {
-            infoDetails.mateValue = convertStringToNumber<int>(tokens.at(++i));
+        } else if ("mate" == token) {            infoDetails.mateValue = convertStringToNumber<int>(tokens.at(++i));
         } else if ("pv" == token) {
             i++;  // skip "pv"
-            for (; i < static_cast<int>(tokens.size()); i++) {
-                infoDetails.pvHalfMoves.emplace_back(tokens.at(i));
+            for (; i < static_cast<int>(tokens.size()); i++) {                infoDetails.pvHalfMoves.emplace_back(tokens.at(i));
             }
         }
     }
@@ -73,30 +71,27 @@ UciInterpreter::InfoDetails UciInterpreter::createInfoDetailsFromInfoTokens(stri
 void UciInterpreter::saveCommonParametersOfBestLine(InfoDetails const& infoDetails) {
     for (StringPair const& nameAndValuePair : infoDetails.commonParameterNameAndValue) {
         if (nameAndValuePair.first == "depth") {
-            m_calculationDetails.depthInPlies = convertStringToNumber<unsigned int>(nameAndValuePair.second);
+            m_calculationDetails.depthInPlies = convertStringToNumber<int>(nameAndValuePair.second);
         } else if (nameAndValuePair.first == "seldepth") {
-            m_calculationDetails.selectiveDepthInPlies = convertStringToNumber<unsigned int>(nameAndValuePair.second);
+            m_calculationDetails.selectiveDepthInPlies = convertStringToNumber<int>(nameAndValuePair.second);
         }
     }
 }
-
 void UciInterpreter::saveVariation(InfoDetails const& infoDetails) {
     if (infoDetails.multipv > 0 && !infoDetails.pvHalfMoves.empty()) {
         auto size = m_calculationDetails.variations.size();
         auto possibleNewSize = infoDetails.multipv;
         auto index = infoDetails.multipv - 1;
         Variation variation{infoDetails.mateValue, infoDetails.scoreInCentipawns, infoDetails.pvHalfMoves};
-        if (possibleNewSize <= size) {
+        if (possibleNewSize <= static_cast<int>(size)) {
             m_calculationDetails.variations[index] = variation;
-        } else if (possibleNewSize == size + 1) {
+        } else if (possibleNewSize == static_cast<int>(size) + 1) {
             m_calculationDetails.variations.emplace_back(variation);
         } else {
-            m_calculationDetails.variations.resize(possibleNewSize);
-            m_calculationDetails.variations[index] = variation;
+            m_calculationDetails.variations.resize(possibleNewSize);            m_calculationDetails.variations[index] = variation;
         }
     }
 }
-
 bool UciInterpreter::shouldSkipTheEntireInfo(string const& token) {
     static const strings tokens{"currmove"};
